+++ forui/test/src/widgets/select/single/select_golden_test.dart
@@ -114,7 +114,7 @@ void main() {
             key: key,
             format: (s) => s,
             style: TestScaffold.blueScreen.selectStyle.copyWith(
-              searchStyle: const .delta(fieldStyle: .delta(cursorColor: Color(0xFF03A9F4))),
+              searchStyle: .delta(fieldStyles: .delta([.all(const .delta(cursorColor: Color(0xFF03A9F4)))])),
             ),
             contentScrollHandles: true,
             filter: (_) => [],
@@ -136,7 +136,7 @@ void main() {
             key: key,
             format: (s) => s,
             style: TestScaffold.blueScreen.selectStyle.copyWith(
-              searchStyle: const .delta(fieldStyle: .delta(cursorColor: Color(0xFF03A9F4))),
+              searchStyle: .delta(fieldStyles: .delta([.all(const .delta(cursorColor: Color(0xFF03A9F4)))])),
             ),
             contentScrollHandles: true,
             filter: (_) async {
@@ -161,7 +161,7 @@ void main() {
             items: {for (int i = 0; i < 10; i++) '$i': i},
             key: key,
             style: TestScaffold.blueScreen.selectStyle.copyWith(
-              searchStyle: const .delta(fieldStyle: .delta(cursorColor: Color(0xFF03A9F4))),
+              searchStyle: .delta(fieldStyles: .delta([.all(const .delta(cursorColor: Color(0xFF03A9F4)))])),
             ),
             contentScrollHandles: true,
             filter: (_) => [],